+++ src/extension.ts
@@ -1,45 +1,76 @@
 // src/extension.ts
-import * as vscode from 'vscode';
-import { ComponentCompletionItemProvider } from './ComponentCompletionItemProvider';
-import { ComponentDefinitionProvider } from './ComponentDefinitionProvider';
-import { refreshComponentIndex } from './ComponentIndexer';
-import { ComponentHoverProvider } from './ComponentHoverProvider';
+import * as vscode from "vscode";
+import { ComponentCompletionItemProvider } from "./ComponentCompletionItemProvider";
+import { ComponentPropsCompletionItemProvider } from "./ComponentPropsCompletionItemProvider";
+import { ComponentDefinitionProvider } from "./ComponentDefinitionProvider";
+import { refreshComponentIndex } from "./ComponentIndexer";
+import { ComponentHoverProvider } from "./ComponentHoverProvider";
+import { ComponentSlotsCompletionItemProvider } from "./ComponentSlotsCompletionItemProvider";
 
 export function activate(context: vscode.ExtensionContext) {
+  refreshComponentIndex();
   const completionProvider = vscode.languages.registerCompletionItemProvider(
-    { language: 'twig', scheme: 'file' },
+    { language: "twig", scheme: "file" },
     new ComponentCompletionItemProvider(),
-    ':', '\'', '"', '/', // Trigger characters
+    ":",
+    "'",
+    '"',
+    "/",
+  );
+
+  const propsCompletionProvider = vscode.languages.registerCompletionItemProvider(
+    { language: "twig", scheme: "file" },
+    new ComponentPropsCompletionItemProvider(),
+    "{",
+    " ",
+    ",",
+    "\n",
+  );
+
+  const slotsCompletionProvider = vscode.languages.registerCompletionItemProvider(
+    { language: "twig", scheme: "file" },
+    new ComponentSlotsCompletionItemProvider(),
+    " ",
   );
 
   const definitionProvider = vscode.languages.registerDefinitionProvider(
-    { language: 'twig', scheme: 'file' },
+    { language: "twig", scheme: "file" },
     new ComponentDefinitionProvider(),
   );
 
   const hoverProvider = vscode.languages.registerHoverProvider(
-    { language: 'twig', scheme: 'file' },
+    { language: "twig", scheme: "file" },
     new ComponentHoverProvider(),
   );
 
-  context.subscriptions.push(completionProvider, definitionProvider, hoverProvider);
+  const watcher = vscode.workspace.createFileSystemWatcher("**/*.{twig,component.yml}");
 
-  // Watch for changes in component files
-  const watcher = vscode.workspace.createFileSystemWatcher('**/*.twig');
+  let debounceTimeout: NodeJS.Timeout;
+  const debouncedRefresh = () => {
+    clearTimeout(debounceTimeout);
+    debounceTimeout = setTimeout(() => {
+      refreshComponentIndex();
+    }, 500);
+  };
 
-  watcher.onDidChange(refreshComponentIndex);
-  watcher.onDidCreate(refreshComponentIndex);
-  watcher.onDidDelete(refreshComponentIndex);
+  watcher.onDidChange(debouncedRefresh);
+  watcher.onDidCreate(debouncedRefresh);
+  watcher.onDidDelete(debouncedRefresh);
 
-  context.subscriptions.push(watcher);
-
-  // Register the refresh index command
-  const refreshCommand = vscode.commands.registerCommand('drupalSDCHelper.refreshIndex', () => {
+  const refreshCommand = vscode.commands.registerCommand("drupalSDCHelper.refreshIndex", () => {
+    vscode.window.showInformationMessage("Refreshing Drupal SDC Index...");
     refreshComponentIndex();
-    vscode.window.showInformationMessage('Drupal SDC component index refreshed.');
   });
 
-  context.subscriptions.push(refreshCommand);
+  context.subscriptions.push(
+    completionProvider,
+    propsCompletionProvider,
+    slotsCompletionProvider,
+    definitionProvider,
+    hoverProvider,
+    watcher,
+    refreshCommand,
+  );
 }
 
 export function deactivate() {